requirements.txt
confluent-kafka
fastavro
pandas
pymysql